refactor(cli): split operations.ts into cohesive core modules
Mechanically split the ~2600-line core/operations.ts into six cohesive
modules under core/:

- shared.ts       — locale/layer lookup helpers (findLocaleImpl,
                    findLocaleOrThrow, findLayerOrThrow, localeRefInfo)
- report.ts       — report-file path validation and resolution
- ops-translate.ts — translateMissing/translateKey, prompt builders,
                    placeholder validation, fallback context,
                    extractJsonFromResponse, protected-locale helpers
- ops-read.ts     — detectConfig, listLocaleDirs, getTranslations,
                    getMissingTranslations, findEmptyTranslations,
                    searchTranslations, listNamespaces
- ops-write.ts    — applyTranslations, write/add/update/remove/rename,
                    scaffoldLocaleFiles
- ops-orphans.ts  — findOrphanKeys, removeOrphanKeys, scanCodeUsage,
                    resolveOrphanScanContext, ignore-pattern helpers

operations.ts remains as a barrel re-exporting the exact historical
export surface, so no consumer (commands/, mcp package, tests,
index.ts) changes. Pure code motion: no behavior changes, no renames,
no signature changes.

Closes #179
